SetupPayload: Python implementation to parse QRCode
* SetupPayload: Python implementation to parse qrcode
* Complete base38 method
* Added parse_setup_payload.py cli command
* Refactor SetupPayload class to a separate, reuseable python class so
  that can be imported by both parse_setup_payload.py and
  generate_setup_payload.py
* Added parse_qrcode static method to SetupPayload class
* Added print method to SetupPayload class to print parsed fields
